+++ src/main/java/roomescape/controller/request/CreateReservationTimeWebRequest.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.web.request;
+package roomescape.controller.request;
 
 import java.time.LocalTime;
 import jakarta.validation.constraints.NotNull;
