+++ src/main/java/roomescape/controller/response/ReservationTimeWebResponse.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.web.response;
+package roomescape.controller.response;
 
 import java.time.LocalTime;
 import com.fasterxml.jackson.annotation.JsonFormat;
